pyproject.toml
[tool.poetry]
name = "template-crewai-langchain"
version = "0.1.0"
description = ""
authors = ["Jimmy Briggs <jimmy.briggs@jimbrig.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
langchain = "^0.1.14"
crewai = "^0.28.0"
requests = "^2.31.0"
pydantic = "^2.6.4"
load-dotenv = "^0.1.0"
crewai-tools = "^0.1.4"
flask = "^3.0.2"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
